# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api",
    "paths": {
...
        "/product/all": {
            "get": {
                "description": "Get all products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ProductsRespOKDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
        "/product/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Create new product",
                "parameters": [
                    {
                        "description": "product data",
                        "name": "product_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.NewProductDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ProductRespOKDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ValidationRespErrDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.RespErrDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
        "/product/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Delete product",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3afc3021-9395-11ed-a8b6-d8bbc1a27045",
                        "description": "product   uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.RespOKDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
        "/product/get/{id}": {
            "get": {
                "description": "Get product by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get product",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3afc3021-9395-11ed-a8b6-d8bbc1a27045",
                        "description": "product   uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ProductRespOKDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.RespErrDTO"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
        "/product/update/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Update product",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3afc3021-9395-11ed-a8b6-d8bbc1a27045",
                        "description": "product   uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "product data",
                        "name": "product_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.UpdateProductDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ProductRespOKDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ValidationRespErrDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.RespErrDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "the token is expired"
                },
                "message": {
                    "type": "string",
                    "example": "invalid access token"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "error querying the database"
                },
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.NewProductDTO": {
            "type": "object",
            "required": [
                "description",
                "images_url",
                "name",
                "price",
                "tags"
            ],
            "properties": {
                "avalible": {
                    "type": "boolean"
                },
                "category": {
                    "type": "string",
                    "example": "clothes"
                },
                "description": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 4,
                    "example": "The best T-shirt in the world."
                },
                "discount_rate": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 23
                },
                "images_url": {
                    "type": "array",
                    "maxItems": 10,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://example.com/image1.png",
                        "https://example.com/image2.png"
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 4,
                    "example": "Black T-Shirt Addidas"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2599
                },
                "tags": {
                    "type": "array",
                    "maxItems": 6,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "t-shirts",
                        "clothes",
                        "addidas"
                    ]
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.ProductDTO": {
            "type": "object",
            "required": [
                "description",
                "images_url",
                "name",
                "price",
                "tags"
            ],
            "properties": {
                "avalible": {
                    "type": "boolean"
                },
                "category": {
                    "type": "string",
                    "example": "clothes"
                },
                "created_at": {
                    "type": "integer",
                    "example": 1674405183
                },
                "description": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 4,
                    "example": "The best T-shirt in the world."
                },
                "discount_rate": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 23
                },
                "id": {
                    "type": "string",
                    "example": "8ded83fe-93c8-11ed-ab0f-d8bbc1a27048"
                },
                "images_url": {
                    "type": "array",
                    "maxItems": 10,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://example.com/image1.png",
                        "https://example.com/image2.png"
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 4,
                    "example": "Black T-Shirt Addidas"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2599
                },
                "tags": {
                    "type": "array",
                    "maxItems": 6,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "t-shirts",
                        "clothes",
                        "addidas"
                    ]
                },
                "updated_at": {
                    "type": "integer",
                    "example": 1674405181
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.ProductRespOKDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ProductDTO"
                },
                "message": {
                    "type": "string",
                    "example": "Data retrived!"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.ProductsRespOKDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ProductDTO"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Data retrived!"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.RespErrDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.RespOKDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Data retrived!"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.UpdateProductDTO": {
            "type": "object",
            "properties": {
                "avalible": {
                    "type": "boolean"
                },
                "category": {
                    "type": "string",
                    "example": "clothes"
                },
                "description": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 4,
                    "example": "The best T-shirt in the world."
                },
                "discount_rate": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 23
                },
                "images_url": {
                    "type": "array",
                    "maxItems": 10,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://example.com/image1.png",
                        "https://example.com/image2.png"
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 4,
                    "example": "Black T-Shirt Addidas"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2599
                },
                "tags": {
                    "type": "array",
                    "maxItems": 6,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "t-shirts",
                        "clothes",
                        "addidas"
                    ]
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.ValidationRespErrDTO": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "field": {
                                "type": "string",
                                "example": "Email"
                            },
                            "message": {
                                "type": "string",
                                "example": "invalid email"
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "one or more field are invalid"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "X-Access-Token",
            "in": "header"
        }
    }
}